mqtt/requirements.txt
gevent==1.2.2
paho-mqtt==1.3.1
